Read company info from book
The gnucash-python bindings seem to lack the necessary functionality to
retrieve company data saved in the book. In the past, I had simply hardcoded
this information in the script. This is dumb, of course, and then my customer
asked me to include my EIN on my invoices, and I did not want to commit that to
github.

The solution was to use ctypes to call the necessary gnucash C API functions.